fixed a bug where the mac quickinstaller would hang during setup when the user has over 255 applications in applications folder

+++ eid-mw/_src/eID-QuickInstaller/eID-EZinstaller/sysdiagnost.cpp
@@ -1822,7 +1822,7 @@ bool CSysDiagnost::isSoftwareInstalled (string productName, vector <Soft_INFO>& 
 #elif __APPLE__
       CFDictionaryRef	appsInfoDict;
 
-      UInt8			i = 0;
+      UInt32			i = 0;
       CFArrayRef		itemsArray;
       CFIndex			arrayCount;
       CFStringRef		sAppName;
